Fixed bots AI

+++ Assets/Scripts/Components/GoBased/RandomNumbers.cs
@@ -4,9 +4,9 @@ namespace Scripts.Components.GoBased
 
     public static class RandomNumbers
     {
-        public static int RandomWithTwoNumber(int first, int second)
+        public static float RandomWithTwoNumber(float first, float second)
         {
-            int[] numbers = new int[2] { first, second };
+            float[] numbers = new float[2] { first, second };
             float rand = Random.Range(0f, 100f);
 
             return rand < 50 ? numbers[0] : numbers[1];
